+++ packages/d6-system-2e-fantasy/d6-system-2e-fantasy.mjs
@@ -16,7 +16,7 @@ Hooks.once("ready", () => {
     mechanicIds: ["fantasy-skills-magic"],
     recommendedPrimaryProfile: "second-edition",
     rulesFamily: "d6-system-second-edition",
-    version: "0.1.0-alpha.28"
+    version: "0.1.0-alpha.29"
   });
 });
 //# sourceMappingURL=d6-system-2e-fantasy.mjs.map